Reorg files, handle post command
